Invulnerability
Unit invulnerability has been implemented. After taking damage, units will be temporarily invulnerable for a specified number of ticks, and cannot take damage during this time.

+++ JARVA/src/engine/states/Game.java
@@ -49,7 +49,7 @@ public class Game extends BasicGameState {
 	/* --- Accessor Methods --- */
 	@Override
 	public int getID() { return id; }
-	public static int getTicks() { return (int) Ticks; }
+	public static float getTicks() { return Ticks; }
 	
 	public ArrayList<GameObject> getGameObjects() { return GameObjects; }
